button selector and responsive stylesgst
