--- app/javascript/components/dropdown-new/dropdown-styles.scss
@@ -1,132 +1,0 @@
-@import '~styles/settings.scss';
-
-$height: rem(35px);
-
-.c-dropdown {
-  text-align: left;
-  color: $slate;
-  width: 100%;
-
-  .label {
-    margin-bottom: rem(10px);
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-    font-weight: 400;
-    font-size: rem(13px);
-
-    .info-button {
-      margin-left: rem(7px);
-      height: 15px;
-      width: 15px;
-      min-width: 15px;
-      border-radius: 50%;
-
-      .info-icon {
-        height: 8px;
-        width: 7px;
-      }
-    }
-  }
-
-  .container {
-    position: relative;
-  }
-
-  .selector {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    position: relative;
-    border: solid 1px $border;
-    height: $height;
-    padding: 0 rem(10px);
-    cursor: pointer;
-    color: $slate;
-    z-index: 0;
-
-    .value {
-      position: absolute;
-      top: calc(50% - 10px);
-      bottom: 0;
-      left: rem(10px);
-      right: rem(10px);
-      padding-right: rem(18px);
-      z-index: 0;
-      white-space: nowrap;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      display: inline-block;
-      font-size: rem(13px);
-    }
-
-    input {
-      z-index: 1;
-      background-color: transparent;
-      width: 100%;
-      height: 100%;
-      border: 0;
-      cursor: pointer;
-      color: $slate;
-      font-size: rem(13px);
-
-      &:focus {
-        outline: none;
-      }
-
-      &::placeholder {
-        color: rgba($slate, 0.5);
-      }
-    }
-
-    .arrow {
-      max-width: rem(15px);
-    }
-  }
-
-  .menu {
-    border: solid 1px $border;
-    margin-top: -1px;
-    z-index: 1;
-    max-height: rem(250px);
-    overflow-y: scroll;
-    cursor: pointer;
-    color: rgba($slate, 0.7);
-    font-size: rem(13px);
-    position: absolute;
-    top: $height;
-    width: 100%;
-    background-color: $white;
-    padding: rem(5px) 0;
-
-    .item-wrapper {
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-    }
-
-    .item {
-      width: 100%;
-      padding: rem(8px) rem(10px);
-    }
-
-    .info-button {
-      margin-right: rem(10px);
-      width: rem(16px);
-      max-width: rem(16px);
-      min-width: rem(16px);
-      height: rem(16px);
-      max-height: rem(16px);
-    }
-
-    .highlight {
-      color: $slate;
-    }
-  }
-
-  .is-open {
-    .arrow {
-      transform: rotate(180deg);
-    }
-  }
-}
