+++ app/javascript/components/dropdown-old/dropdown-styles.scss
@@ -1,0 +1,109 @@
+@import '~styles/settings.scss';
+
+.c-dropdown {
+  text-align: left;
+  color: $slate;
+
+  .label {
+    margin-bottom: rem(10px);
+    display: flex;
+    justify-content: flex-start;
+    align-items: center;
+    font-weight: 400;
+    font-size: rem(13px);
+
+    .info-button {
+      margin-left: rem(7px);
+      height: 15px;
+      width: 15px;
+      min-width: 15px;
+      border-radius: 50%;
+
+      .info-icon {
+        height: 8px;
+        width: 7px;
+      }
+    }
+  }
+
+  .dd__selectControl {
+    padding: 0;
+  }
+
+  .dd__selected {
+    padding: rem(5px) rem(5px) rem(5px) rem(10px);
+  }
+
+  .icon {
+    margin-right: rem(10px);
+  }
+
+  .dd__selected,
+  .dd__placeholder,
+  .dd__selectedItem,
+  .dd__list {
+    > div {
+      white-space: nowrap;
+      overflow: hidden;
+      text-overflow: ellipsis;
+      height: 100%;
+    }
+  }
+
+  .dd__search {
+    text-overflow: initial !important;
+  }
+
+  .dd__list {
+    display: block;
+    width: 100%;
+    box-shadow: none;
+    border: solid 1px #e5e5df;
+    border-radius: 0;
+  }
+
+  .dd__option {
+    white-space: normal;
+    width: 100%;
+    color: rgba($slate, 0.6);
+    display: flex;
+    justify-content: space-between;
+    flex-direction: row;
+    align-items: center;
+
+    &:hover {
+      color: rgba($slate, 1);
+      background-color: $white;
+    }
+
+    .info-button {
+      margin-left: rem(10px);
+      height: 15px;
+      width: 15px;
+      min-width: 15px;
+      border-radius: 50%;
+
+      .info-icon {
+        height: 8px;
+        width: 7px;
+      }
+    }
+  }
+
+  .dd__selectedItem,
+  .dd__selectedOption {
+    color: rgba($slate, 1);
+  }
+
+  .dd__selectedOption {
+    background-color: $white;
+  }
+
+  .dd__disabled {
+    .dd__selectControl {
+      cursor: not-allowed;
+      background-color: $grey-light;
+      border-color: $grey-dark;
+    }
+  }
+}
