+++ app/javascript/components/dropdown/themes/dropdown-button.scss
@@ -1,0 +1,51 @@
+@import '~styles/settings.scss';
+
+$dd-button-height: rem(28px);
+
+.theme-dropdown-button {
+  text-align: center;
+
+  .container {
+    max-width: rem(85px);
+
+    .selector {
+      border-radius: rem(25px);
+      border-color: $green-gfw;
+      height: $dd-button-height;
+
+      .value {
+        padding: 0;
+      }
+    }
+
+    &.is-open {
+      .menu-arrow {
+        position: absolute;
+        width: 0;
+        height: 0;
+        margin: 0 9px;
+        border-bottom: 7px solid #fff;
+        border-left: 7px solid transparent;
+        border-left-color: #fff;
+        border-right: none;
+        transform-origin: 0 0;
+        transform: rotate(-45deg);
+        top: calc(#{$dd-button-height} + 6px);
+        left: calc(50% - 14px);
+        border-bottom-color: #fff;
+        box-shadow: 1px -2px 3px 0 rgba(85, 85, 85, 0.1);
+        z-index: 3;
+      }
+    }
+
+    .menu {
+      top: calc(#{$dd-button-height} + 5);
+      border: 0;
+      box-shadow: 1px 1px 10px 0 rgba(85, 85, 85, 0.25);
+    }
+  }
+
+  .arrow {
+    display: none;
+  }
+}
