+++ app/javascript/components/dropdown/themes/dropdown-dark.scss
@@ -1,0 +1,57 @@
+@import '~styles/settings.scss';
+
+$dd-dark-height: rem(70px);
+$dd-dark-font-size: rem(47px);
+
+.theme-dropdown-dark {
+  color: $white;
+  font-weight: 300;
+
+  .selector {
+    border: 0;
+    height: $dd-dark-height;
+    padding: 0;
+    color: $white;
+
+    &.align-left {
+      .value {
+        padding-left: rem(30px);
+      }
+    }
+
+    .value {
+      font-size: $dd-dark-font-size;
+      top: 0;
+      left: 0;
+      right: 0;
+      padding-right: 0;
+    }
+
+    input {
+      color: $white;
+      font-size: $dd-dark-font-size;
+      font-weight: 300;
+      padding-left: rem(18px);
+
+      &::placeholder {
+        color: rgba($white, 0.5);
+      }
+    }
+
+    .arrow {
+      fill: $white;
+    }
+  }
+
+  .menu {
+    border: solid 1px $border;
+    font-size: rem(18px);
+    font-weight: 400;
+    top: $dd-dark-height;
+    box-shadow: 1px 1px 5px 0 rgba(85, 85, 85, 0.25);
+
+    @media screen and (min-width: $screen-m) {
+      left: rem(30px);
+    }
+  }
+}
