+++ app/javascript/pages/country/header/header-styles.scss
@@ -43,20 +43,8 @@
       margin-bottom: 0;
     }
 
-    .select {
-      display: flex;
-      flex-direction: row;
-      justify-content: flex-start;
-      align-items: center;
-      margin-bottom: rem(20px);
-
-      .icon {
-        min-width: rem(12px);
-        width: rem(12px);
-        height: rem(12px);
-        margin-right: rem(16px);
-        fill: $white;
-      }
+    > div {
+      margin-bottom: rem(5px);
     }
   }
